メソッド分割
